experiments/pro_exp/requirements.txt
bert-extractive-summarizer
Cython
deepspeech
fastapi
ffmpeg
future
ipywidgets
nemo_toolkit[all]
neuralcoref
pynini
pytube
scipy
sentencepiece
spacy
torch
torchvision
transformers
uvicorn
wave
